test(back): #919 add tests
- Add a few unit and functinoal tests
