+++ .github/workflows/dev.yml
@@ -501,6 +501,23 @@ jobs:
       - name: /taintTerraform/module
         run: nix-env -if . && m . /taintTerraform/module
 
+  linux_testPython_cliMain:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@2541b1294d2704b0964813337f33b291d3f8596b
+      - uses: docker://docker.io/nixos/nix@sha256:1d13ae379fb8caf3f859c5ce7ec6002643d60cf8b7b6147b949cc34880c93bac
+        name: /testPython/cliMain
+        with:
+          set-safe-directory: /github/workspace
+          args: sh -c "nix-env -if . && m . /testPython/cliMain"
+  macos_testPython_cliMain:
+    runs-on: macos-latest
+    steps:
+      - uses: actions/checkout@2541b1294d2704b0964813337f33b291d3f8596b
+      - uses: cachix/install-nix-action@451e61183802597c1febd6ca3cf18aa163f93a06
+      - name: /testPython/cliMain
+        run: nix-env -if . && m . /testPython/cliMain
+
   linux_testPython_example:
     runs-on: ubuntu-latest
     steps:
