+++ .github/workflows/prod.yml
@@ -98,6 +98,27 @@ jobs:
         env:
           CACHIX_AUTH_TOKEN: ${{ secrets.CACHIX_AUTH_TOKEN }}
 
+  linux_dev_cliMain:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@2541b1294d2704b0964813337f33b291d3f8596b
+      - uses: docker://docker.io/nixos/nix@sha256:1d13ae379fb8caf3f859c5ce7ec6002643d60cf8b7b6147b949cc34880c93bac
+        name: /dev/cliMain
+        with:
+          set-safe-directory: /github/workspace
+          args: sh -c "nix-env -if . && m . /dev/cliMain"
+        env:
+          CACHIX_AUTH_TOKEN: ${{ secrets.CACHIX_AUTH_TOKEN }}
+  macos_dev_cliMain:
+    runs-on: macos-latest
+    steps:
+      - uses: actions/checkout@2541b1294d2704b0964813337f33b291d3f8596b
+      - uses: cachix/install-nix-action@451e61183802597c1febd6ca3cf18aa163f93a06
+      - name: /dev/cliMain
+        run: nix-env -if . && m . /dev/cliMain
+        env:
+          CACHIX_AUTH_TOKEN: ${{ secrets.CACHIX_AUTH_TOKEN }}
+
   linux_dev_example:
     runs-on: ubuntu-latest
     steps:
