+++ src/cli/main/emojis.py
@@ -1,0 +1,94 @@
+# SPDX-FileCopyrightText: 2022 Fluid Attacks and Makes contributors
+#
+# SPDX-License-Identifier: MIT
+
+from typing import (
+    List,
+)
+
+FAILURE: List[str] = [
+    "alien_monster",  # 👾
+    "anxious_face_with_sweat",  # 😰
+    "beetle",  # 🐞
+    "blowfish",  # 🐡
+    "brick",  # 🧱
+    "broken_heart",  # 💔
+    "bug",  # 🐛
+    "collision",  # 💥
+    "dizzy_face",  # 😵
+    "exploding_head",  # 🤯
+    "eyes",  # 👀
+    "face_with_monocle",  # 🧐
+    "fire",  # 🔥
+    "ghost",  # 👻
+    "lady_beetle",  # 🐞
+    "mega",  # 📣
+    "microscope",  # 🔬
+    "moai",  # 🗿
+    "open_mouth",  # 😮
+    "person_facepalming",  # 🤦
+    "person_getting_massage",  # 💆
+    "sad_but_relieved_face",  # 😥
+    "see_no_evil",  # 🙈
+    "smiling_imp",  # 😈
+    "speak_no_evil",  # 🙊
+    "thinking_face",  # 🤔
+    "upside__down_face",  # 🙃
+    "volcano",  # 🌋
+    "wilted_flower",  # 🥀
+    "woozy_face",  # 🥴
+    "yawning_face",  # 🥱
+    "zipper__mouth_face",  # 🤐
+]
+SUCCESS: List[str] = [
+    "airplane_departure",  # 🛫
+    "beer",  # 🍺
+    "beers",  # 🍻
+    "birthday",  # 🎂
+    "bottle_with_popping_cork",  # 🍾
+    "bouquet",  # 💐
+    "bulb",  # 💡
+    "blossom",  # 🌼
+    "boxing_glove",  # 🥊
+    "call_me_hand",  # 🤙
+    "cat",  # 🐱
+    "clapping_hands",  # 👏
+    "clinking_glasses",  # 🥂
+    "colombia",  # 🇨🇴
+    "confetti_ball",  # 🎊
+    "couple_with_heart",  # 💑
+    "checkered_flag",  # 🏁
+    "crown",  # 👑
+    "dart",  # 🎯
+    "dog",  # 🐶
+    "dancer",  # 💃
+    "doughnut",  # 🍩
+    "eagle",  # 🦅
+    "elephant",  # 🐘
+    "face_blowing_a_kiss",  # 😘
+    "flamingo",  # 🦩
+    "four_leaf_clover",  # 🍀
+    "fries",  # 🍟
+    "glowing_star",  # 🌟
+    "kite",  # 🪁
+    "mage",  # 🧙
+    "merperson",  # 🧜
+    "money_with_wings",  # 💸
+    "nail_care",  # 💅
+    "party_popper",  # 🎉
+    "partying_face",  # 🥳
+    "person_cartwheeling",  # 🤸
+    "person_playing_handball",  # 🤾
+    "person_playing_water_polo",  # 🤽
+    "person_surfing",  # 🏄
+    "pizza",  # 🍕
+    "popcorn",  # 🍿
+    "rainbow",  # 🌈
+    "shooting_star",  # 🌠
+    "smiling_face_with_sunglasses",  # 😎
+    "smirk",  # 😏
+    "rocket",  # 🚀
+    "trophy",  # 🏆
+    "whale",  # 🐳
+    "wink",  # 😉
+]
